refactor: remove unecessary intermediate state
