+++ packages/manager/modules/cda/src/details/cda-details.html
@@ -6,25 +6,25 @@
     >
         <oui-header-tabs>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-details.cda-details-home"
+                state="cda.dashboard.cda-details-home"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_home"></span>
             </oui-header-tabs-item>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-details.cda-user.cda-user-list"
+                state="cda.dashboard.cda-user.cda-user-list"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_user"></span>
             </oui-header-tabs-item>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-details.cda-ip.cda-ip-list"
+                state="cda.dashboard.cda-ip.cda-ip-list"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_ip"></span>
             </oui-header-tabs-item>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-details.cda-pool.cda-pool-list"
+                state="cda.dashboard.cda-pool.cda-pool-list"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_pool"></span>
