+++ packages/manager/modules/cda/src/details/cda.constants.js
