+++ packages/manager/modules/cda/src/details/cda.less
