+++ packages/manager/modules/cda/src/details/cda.service.js
