+++ packages/manager/modules/cda/src/details/home/cda-details-home.js
@@ -2,7 +2,7 @@ import template from './cda-details-home.html';
 import controller from './cda-details-home.controller';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-details-home', {
+  $stateProvider.state('cda.dashboard.cda-details-home', {
     url: '/home',
     views: {
       cdaDetailsTab: {
