+++ packages/manager/modules/cda/src/details/translations/Messages_de_DE.json
