+++ packages/manager/modules/cda/src/details/translations/Messages_fr_CA.json
