+++ packages/manager/modules/cda/src/details/translations/Messages_fr_FR.json
