+++ packages/manager/modules/cda/src/details/translations/Messages_pl_PL.json
