+++ packages/manager/modules/cda/src/details/translations/Messages_pt_PT.json
