+++ packages/manager/modules/cda/src/ip/cda-ip.js
@@ -1,7 +1,7 @@
 import template from './cda-ip.html';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-ip', {
+  $stateProvider.state('cda.dashboard.cda-ip', {
     url: '/ip',
     views: {
       cdaDetailsTab: {
