+++ packages/manager/modules/cda/src/ip/list/cda-ip-list.js
@@ -2,7 +2,7 @@ import controller from './cda-ip-list.controller';
 import template from './cda-ip-list.html';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-ip.cda-ip-list', {
+  $stateProvider.state('cda.dashboard.cda-ip.cda-ip-list', {
     url: '/list',
     views: {
       cdaIpContent: {
