+++ packages/manager/modules/cda/src/pool/cda-pool.js
@@ -1,7 +1,7 @@
 import template from './cda-pool.html';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-pool', {
+  $stateProvider.state('cda.dashboard.cda-pool', {
     url: '/pool',
     views: {
       cdaDetailsTab: {
