+++ packages/manager/modules/cda/src/pool/list/cda-pool-list.js
@@ -2,7 +2,7 @@ import controller from './cda-pool-list.controller';
 import template from './cda-pool-list.html';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-pool.cda-pool-list', {
+  $stateProvider.state('cda.dashboard.cda-pool.cda-pool-list', {
     url: '/list',
     views: {
       cdaPoolContent: {
