+++ packages/manager/modules/cda/src/user/cda-user.js
@@ -1,7 +1,7 @@
 import template from './cda-user.html';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-user', {
+  $stateProvider.state('cda.dashboard.cda-user', {
     url: '/user',
     views: {
       cdaDetailsTab: {
