+++ packages/manager/modules/cda/src/user/details/cda-user-details.html
@@ -1,7 +1,7 @@
 <section class="cda-user-details">
     <a
         class="oui-button oui-button_icon-left oui-button_ghost"
-        data-ui-sref="cda.dashboard.cda-details.cda-user.cda-user-list"
+        data-ui-sref="cda.dashboard.cda-user.cda-user-list"
     >
         <i class="oui-icon oui-icon-chevron-left" aria-hidden="true"></i>
         {{ "cda_user_details_btn_back" | translate }}
