+++ packages/manager/modules/cda/src/user/details/permissions/cda-user-details-permission-list.html
@@ -5,7 +5,7 @@
     <a
         class="oui-button oui-button_secondary"
         data-ng-if="!CdaUserDetailsPermissionListCtrl.loading"
-        data-ui-sref="cda.dashboard.cda-details.cda-user.cda-user-details.cda-user-details-permission-list.cda-user-details-permission-list-edit"
+        data-ui-sref="cda.dashboard.cda-user.cda-user-details.cda-user-details-permission-list.cda-user-details-permission-list-edit"
         role="button"
     >
         <span
