+++ packages/manager/modules/cda/src/user/details/permissions/edit/cda-user-details-permission-list-edit.html
@@ -1,7 +1,7 @@
 <section class="cda-user-details-permission-list-edit cui-page__content">
     <a
         class="oui-button oui-button_icon-left oui-button_ghost"
-        data-ui-sref="cda.dashboard.cda-details.cda-user.cda-user-details.cda-user-details-permission-list({serviceName: CdaUserDetailsPermissionListEditCtrl.serviceName, userName: CdaUserDetailsPermissionListEditCtrl.userName})"
+        data-ui-sref="cda.dashboard.cda-user.cda-user-details.cda-user-details-permission-list({serviceName: CdaUserDetailsPermissionListEditCtrl.serviceName, userName: CdaUserDetailsPermissionListEditCtrl.userName})"
     >
         <i class="oui-icon oui-icon-chevron-left"></i>
         {{ "cda_user_details_permissions_list_edit_btn_back" | translate }}
