+++ packages/manager/modules/cda/src/user/list/cda-user-list.controller.js
@@ -92,7 +92,7 @@ export default class CdaUserListCtrl {
 
     self.viewPermissions = function viewPermissions(name) {
       return $state.go(
-        'cda.dashboard.cda-details.cda-user.cda-user-details.cda-user-details-permission-list',
+        'cda.dashboard.cda-user.cda-user-details.cda-user-details-permission-list',
         { userName: name },
       );
     };
