+++ packages/manager/modules/cda/src/user/list/cda-user-list.js
@@ -2,7 +2,7 @@ import controller from './cda-user-list.controller';
 import template from './cda-user-list.html';
 
 export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-details.cda-user.cda-user-list', {
+  $stateProvider.state('cda.dashboard.cda-user.cda-user-list', {
     url: '/list',
     views: {
       cdaUserContent: {
